server/requirements.txt
Django==4.2.7
djangorestframework==3.16.0
django-cors-headers==4.0.0
django-filter==23.5
django-extensions==3.2.1
channels==4.0.0
djangorestframework-simplejwt==5.2.2
Pillow==10.4.0
django-phonenumber-field==7.1.0
drf-spectacular==0.26.2
celery==5.3.1
redis==4.5.5
django-celery-beat==2.5.0
requests==2.31.0
sendgrid==6.11.0
stripe==12.5.0
phonenumbers==9.0.13